Rename Devices.Can to Device.Can
