+++ ChibiOS/ST_STM32F4_DISCOVERY/target_nf_device_can_config.h
